feat: add paged job posting filters

+++ jobpilot-ai/backend/src/main/java/com/jobpilot/api/domain/jobposting/dto/JobPostingLocationResponse.java
@@ -1,11 +1,13 @@
 package com.jobpilot.api.domain.jobposting.dto;
 
+import java.math.BigDecimal;
+
 public record JobPostingLocationResponse(
         String locationText,
         String sido,
         String sigungu,
         String detailedAddress,
-        Double latitude,
-        Double longitude,
+        BigDecimal latitude,
+        BigDecimal longitude,
         boolean primaryLocation
 ) {}
